main: Pass the chip ID to Linux
The chip ID register isn't guaranteed to be at a fixed location, so under
some circumstances the bootloader might know it but Linux doesn't.  Pass
its value to the kernel in $a3.

Linux probably doesn't need this information because it can gather
everything it needs from DT, but just in case something comes up in the
future, it will be there.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>