Add experimental "make VIPER=1" build
This generates a bootloader suitable for running Linux on the second
thread of the Viper CPU instead of the Zephyr CPU.

For now, the output image is still called zephyr.img, and switching
between Zephyr/Viper images in the same tree requires an explicit
"make clean".  It isn't clear whether this will ever be used for
production, or just testing purposes.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>