buildsys: Don't use "make dtbs" anymore
Now that Kconfig asks the user to pick a single DT file, we can't just
run "make dtbs" and then grab the one we want.  Also, it is somewhat
clunky for the board.dtb target to depend on having a Linux tree
configured, and there was no way around this.

So, the new behavior is to build board.dtb directly using dtc, from the
dts file in the Linux tree, iff it doesn't already exist.  It will not
list any dependencies.

Signed-off-by: Kevin Cernekee <cernekee@gmail.com>